+++ modules/LinearBoltzmannSolvers/Executors/pi_keigen_scdsa.h
@@ -46,15 +46,28 @@ public:
   void Initialize() override;
   void Execute() override;
 
+  /**
+   * Copies only the scalar moments from an lbs primary flux moments vector.
+   */
   std::vector<double> CopyOnlyPhi0(const LBSGroupset& groupset, const std::vector<double>& phi_in);
 
+  /**
+   * Copies back only the scalar moments to a lbs primary flux vector.
+   */
   void ProjectBackPhi0(const LBSGroupset& groupset,
                        const std::vector<double>& input,
                        std::vector<double>& output);
 
+  /**
+   * Creates a ghost communicator and all associated information.
+   */
   GhostInfo MakePWLDVecGhostCommInfo(const chi_math::SpatialDiscretization& sdm,
                                      const chi_math::UnknownManager& uk_man);
 
+  /**
+   * This method takes an input vector that is the local version of a PWLD discrete space and then
+   * makes it continuous by applying nodal averages.
+   */
   static std::vector<double>
   NodallyAveragedPWLDVector(const std::vector<double>& input,
                             const chi_math::SpatialDiscretization& pwld_sdm,
